+++ static/views/property-meta.php
@@ -9,9 +9,17 @@
 <?php if(is_array( $meta )): ?>
   <?php foreach( $meta as $meta_slug => $meta_title ):
     $meta_value = do_shortcode(html_entity_decode(get_post_meta( $post_id, $meta_slug, true )));
-    if(trim($meta_value) != ""):
+    if( isset($meta_value) &&  trim($meta_value) != ""):
     ?>
-      <h2><?php echo $meta_title; ?></h2>
+        <?php
+        // added : 22/07/2016 @raj
+        // show title only if enabled
+        if(isset($show_taxonomy_label) && $show_taxonomy_label !== "false" ) {
+            echo  '<h2>'. $meta_title .'</h2>';
+        }
+        //end : @raj
+        ?>
+         
       <p><?php echo trim($meta_value);?></p>
     <?php 
     endif;